Deleted other references to people.
What every person did should be in an 'Author contributions' section at the end, not in main text. That is if you think having a specific contribution list stating who made what script is even necessary. 
Also, whether JAFY had computational training isn't really relevant to a reader. Makes JAFY more sound relatable, sure, but it is not of significance for the pipeline.